+++ src/lib.rs
@@ -128,7 +128,7 @@ pub struct Spidev {
 /// Options that are not configured with one of the builder
 /// functions will not be modified in the kernel when
 /// `configure` is called.
-#[derive(Clone)]
+#[derive(Default, Clone)]
 pub struct SpidevOptions {
     pub bits_per_word: Option<u8>,
     pub max_speed_hz: Option<u32>,