+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 [Full Changelog](https://github.com/rust-embedded/rust-spidev/compare/0.4.1...HEAD)
 
+- Added `Default` implementation for `SpidevOptions`
 - Nix bumped to 0.22
 - bitflags updated to 1.3
 - Minimum supported rust version is now 1.46.0